requirements.txt
pillow
qrcode
pandas